LayoutTests/fast/mediastream/RTCIceCandidate-expected.txt

Tests RTCIceCandidate.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS candidate = new RTCIceCandidate(initializer); did not throw exception.
PASS candidate.candidate is "foo"
PASS candidate.sdpMid is "bar"
PASS candidate.sdpMLineIndex is 6

Dictionary member 'candidate' is required.
PASS new RTCIceCandidate({}); threw exception TypeError: Error creating RTCIceCandidate.
PASS new RTCIceCandidate({sdpMid:"bar", sdpMLineIndex:6}); threw exception TypeError: Error creating RTCIceCandidate.
PASS new RTCIceCandidate({sdpMid:"bar"}); threw exception TypeError: Error creating RTCIceCandidate.
PASS new RTCIceCandidate({sdpMLineIndex:6}); threw exception TypeError: Error creating RTCIceCandidate.

One of the 'sdpMid' or 'sdpMLineIndex' members must be present.
PASS new RTCIceCandidate({candidate:"foo"}); threw exception TypeError: Error creating RTCIceCandidate.

When one of the 'sdpMid' or 'sdpMLineIndex' members is set, the other must be null.
PASS candidate = new RTCIceCandidate({candidate:"foo", sdpMid:"bar"}); did not throw exception.
PASS candidate.sdpMid is "bar"
PASS candidate.sdpMLineIndex is null
PASS candidate = new RTCIceCandidate({candidate:"foo", sdpMLineIndex:6}); did not throw exception.
PASS candidate.sdpMid is null
PASS candidate.sdpMLineIndex is 6

Some bad input
PASS new RTCIceCandidate({candidate:"foo", sdpMLineIndex:"a"}); threw exception TypeError: Error creating RTCIceCandidate.
PASS new RTCIceCandidate({candidate:"foo", sdpMLineIndex:-1}); threw exception TypeError: Error creating RTCIceCandidate.
PASS new RTCIceCandidate({candidate:"foo", sdpMLineIndex:65536}); threw exception TypeError: Error creating RTCIceCandidate.
PASS new RTCIceCandidate(5); threw exception TypeError: Optional RTCIceCandidate constructor argument must be a valid Dictionary.
PASS new RTCIceCandidate("foobar"); threw exception TypeError: Optional RTCIceCandidate constructor argument must be a valid Dictionary.
PASS new RTCIceCandidate(candidate:""); threw exception SyntaxError: Unexpected token ':'. Expected ')' to end a argument list..
PASS successfullyParsed is true

TEST COMPLETE
